feat: 네트워크가 연결되어 있지 않으면 얼럿 띄우도록 구현
