--- ProjectManager/ProjectManager/NetworkCheckManager.swift
@@ -1,49 +1,0 @@
-import Foundation
-import Network
-
-class NetworkCheckManager {
-    private let monitor: NWPathMonitor
-    @Published private(set) var isConnected: Bool = false
-    @Published private(set) var connectionType: ConnectionType = .unknown
-    
-    enum ConnectionType {
-        case wifi
-        case cellular
-        case ethernet
-        case unknown
-    }
-    
-    private init() {
-        monitor = NWPathMonitor()
-    }
-    
-    public func startMonitoring() {
-        monitor.start(queue: DispatchQueue.global())
-        monitor.pathUpdateHandler = { [weak self] path in
-            self?.isConnected = path.status == .satisfied
-            self?.getConnectionType(path)
-            
-            if self?.isConnected == true {
-                print("연결됨")
-            } else {
-                print("연결안됨")
-            }
-        }
-    }
-    
-    public func stopMonitoring() {
-        monitor.cancel()
-    }
-    
-    private func getConnectionType(_ path: NWPath) {
-        if path.usesInterfaceType(.wifi) {
-            connectionType = .wifi
-        } else if path.usesInterfaceType(.cellular) {
-            connectionType = .cellular
-        } else if path.usesInterfaceType(.wiredEthernet) {
-            connectionType = .ethernet
-        } else {
-            connectionType = .unknown
-        }
-    }
-}
